pyproject.toml
[tool.poetry]
name = "api-test"
version = "0.1.0"
description = ""
authors = ["Elena Donyakina <hellengoll@gmail.com>"]
readme = "README.md"
packages = [{include = "api_test"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
allure-pytest = "^2.13.2"
pytest = "^7.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
